+++ spec/features/maths_and_physics_claim_spec.rb
@@ -133,6 +133,7 @@ RSpec.feature "Maths & Physics claims" do
       expect(page).to have_text("Check your answers before sending your application")
 
       stub_geckoboard_dataset_update
+      stub_qualified_teaching_status_show(claim: claim)
 
       freeze_time do
         perform_enqueued_jobs do