+++ spec/lib/dqt/client_spec.rb
@@ -111,7 +111,8 @@ module Dqt
 
     let(:client_args) do
       {
-        host: "test"
+        host: "test",
+        port: nil
       }
     end